change return to float

+++ divider.go
@@ -1,5 +1,5 @@
 package gonumberdivider
 
-func Divide(a, b int) int {
-	return a / b
+func Divide(a, b int) float64 {
+	return float64(a) / float64(b)
 }
